backend/requirements.txt
# Core Django
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Django Extensions
django-cors-headers==4.9.0
django-filter==25.2
django-extensions==4.1
django-widget-tweaks==1.5.0
dj-rest-auth==7.0.1
django-allauth==65.10.0
drf-spectacular==0.29.0

# Database
psycopg2-binary==2.9.11

# Channels (WebSocket)
channels==4.3.1
channels-redis==4.3.0
daphne==4.2.1

# Redis
redis==7.0.1

# Environment Variables
python-dotenv==1.2.1

# Google AI (Gemini)
google-generativeai==0.8.3
google-api-python-client==2.149.0
google-auth==2.36.0
google-auth-httplib2==0.2.0

# Machine Learning & NLP
sentence-transformers==5.1.2
scikit-learn==1.7.2
transformers==4.57.1
torch==2.9.1
numpy>=2.0.0
pandas>=2.2.0

# Utilities
Pillow==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
PyJWT==2.10.1
Faker==37.4.2

# Payment Processing
stripe>=7.0.0bleach==6.3.0
